AlertDialog refactor
